KPMP-205: Handle saving of file

+++ src/main/java/org/kpmp/packages/Attachment.java
@@ -36,7 +36,7 @@ public class Attachment {
 		return size;
 	}
 
-	public void setSize(int size) {
+	public void setSize(long size) {
 		this.size = size;
 	}
